[Issue-751] Use the standard WildFly SECURITY.md for this repo; use a different file name for the content for the rendered security reporting page.
It seems if Github sees a security.md file it regards it as declaring the repo's security policy. But we have a file with that name that doesn't do that; it just provides some minimal jekyll front matter used for the rendered https://wildfly.org/security/ page.

So, rename that file and provide a proper SECURITY.md.